requirements.txt
Flask==2.0.2
Flask_Login==0.5.0
Flask_SQLAlchemy==2.5.1
itsdangerous==2.0.1
opencv_python==4.5.5.62
pandas==1.4.1
Pillow==9.0.1
scikit_image==0.19.2
scipy==1.8.0
torch==1.11.0
torchvision==0.12.0
Werkzeug==2.0.3
requests~=2.27.1
pymongo~=4.1.1
